backend/requirements.txt
fastapi==0.112.0
fastapi-cors==0.0.6
uvicorn==0.30.5
python-dotenv==1.0.1
google-cloud-aiplatform==1.60.0
google-cloud-storage==2.18.0
python-multipart==0.0.9
black==24.3.0
cffi==1.15.1
cryptography==42.0.4
flake8==6.0.0
isort==5.11.4
pre-commit==2.21.0
pycparser==2.21
PyJWT==2.4.0
secure==0.3.0
pydantic-settings==2.4.0
google-generativeai==0.7.2
starlette~=0.37.2
pydantic~=2.8.2